Add rule description xaml builder smoke test for rich rule descriptions (#5011)
Fixes #4601
